pokedex/dagger_python/requirements.txt
pip~=24.0
dagger-io==0.14.0
anyio~=4.3.0